python/opensea-buy/requirements.txt
axonfi>=0.2.0
requests>=2.31.0
python-dotenv>=1.0.0
eth-abi>=5.0.0
